rename loading component
