+++ src/pages/tasks/TaskOverview.page.vue
@@ -1,6 +1,6 @@
 <template>
 	<DefaultWireframe :headline="t('common.words.tasks')" max-width="limited" :fab-items="fabItems">
-		<SvsLoadingWrapper :is-loading="isLoadingTasks && !hasLoadedOnce">
+		<SvsLoading :is-loading="isLoadingTasks && !hasLoadedOnce">
 			<template #loading>
 				<div class="d-flex flex-column w-100">
 					<VSkeletonLoader type="text" :max-width="'15%'" />
@@ -11,7 +11,7 @@
 				<TasksOverviewStudent v-if="isStudent" />
 				<TasksOverviewTeacher v-else-if="isTeacher" />
 			</template>
-		</SvsLoadingWrapper>
+		</SvsLoading>
 	</DefaultWireframe>
 </template>
 
@@ -23,7 +23,7 @@ import { Permission } from "@api-server";
 import { useAppStore, useAppStoreRefs } from "@data-app";
 import { useTasksOfOverview } from "@data-tasks";
 import { mdiPlus } from "@icons/material";
-import { SvsLoadingWrapper } from "@ui-containers";
+import { SvsLoading } from "@ui-containers";
 import { DefaultWireframe } from "@ui-layout";
 import { useTitle } from "@vueuse/core";
 import { computed, ref, watch } from "vue";
