+++ src/pages/user-login-migration/UserLoginMigrationConsent.page.vue
@@ -1,5 +1,5 @@
 <template>
-	<SvsLoadingWrapper :is-loading="isLoading">
+	<SvsLoading :is-loading="isLoading">
 		<div class="text-center mx-auto container-max-width">
 			<img src="@/assets/img/migration/move.svg" alt="" />
 			<h1 class="px-4">
@@ -51,14 +51,14 @@
 				</div>
 			</div>
 		</div>
-	</SvsLoadingWrapper>
+	</SvsLoading>
 </template>
 
 <script setup lang="ts">
 import { buildPageTitle } from "@/utils/pageTitle";
 import { useSystem } from "@data-access";
 import { useUserLoginMigration } from "@data-user-login-migration";
-import { SvsLoadingWrapper } from "@ui-containers";
+import { SvsLoading } from "@ui-containers";
 import { useTitle } from "@vueuse/core";
 import { computed, onMounted, ref } from "vue";
 import { useI18n } from "vue-i18n";
